data_orchestrator/requirements.txt
# Azure Functions requirements
azure-functions>=1.15.0
azure-storage-blob>=12.19.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3 